LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/panner-distance-clamping-expected.txt

CONSOLE MESSAGE: line 67: ReferenceError: Can't find variable: PannerNode

Harness Error (FAIL), message = ReferenceError: Can't find variable: PannerNode

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [ref-distance-error]  
FAIL X new PannerNode(c, {refDistance: -1}) threw "ReferenceError" instead of RangeError. assert_true: expected true got false
FAIL X new PannerNode(c, {refDistance: 0}) incorrectly threw ReferenceError: "Can't find variable: PannerNode". assert_true: expected true got false
FAIL X new PannerNode(c, {refDistance: 5e-324}) incorrectly threw ReferenceError: "Can't find variable: PannerNode". assert_true: expected true got false
